Upgrade to Lebab 2.0
Lebab 2.0 exposes a different programming API: instead of
needing to instantiate Transformer class with object that specifies
the transforms, one now simply uses lebab.transform() which takes
an array of transform names as configuration.

Lebab 2.0 also adds an additional exponent transform, but I'm
not enabling this as it's an ES7 transform.