Replace convert() method with a class
Move all the helper functions into the class as methods.

Leave the try-catch loop outside the class - this seems
to be some safety-net, not really part of the conversion logic.

That's all a preparation to allow us to replace various
variables with object fields, so they can be more easily
shared between methods.